irc-im: fixed invalid memory reading on chat leave
When a chat is left, prpl->chat_leave() is invoked, which is suppose
to free the groupchat. Since the data is now freed, or suppose to have
been freed, attempting to modify the data will result in bad things.
This simply removes an assignment operation, which was unneeded due to
the memory already being freed.