bee-chat: create temporary users for unknown chat participants
The imcb_chat_msg() function is unable to send messages to a chat with
a user who was not previously added. This function should allow for the
sending of messages with users who are not added. This is suitable for
protocols which are sending messages to a chat from random users or a
large amount of users which join and part frequently.

Conflicts:
	protocols/bee_chat.c